Active abilities performance improvements and new map elements
[•] Refactored all active abilities, improving performance for several of them
[•] Replaced Arcane Pulse sprite with a more detailed version
[•] Added "Fountain of Wisdom" to the map
[•] Implemented damage popups
[•] Fixed font-related issues
[•] Fixed unlockables item border color not updating when switching categories or exiting the unlockables menu
[•] Implemented experience crystal scaling

+++ Assets/_Scripts/GameObjects/Player.cs
@@ -259,7 +259,7 @@ public class Player : MonoBehaviour
     /// </summary>
     /// <param name="experienceSource">Origin of the experience.</param>
     /// <param name="experienceToAdd">Amount of experience to add.</param>
-    public void AddExperience(UnityEngine.Object experienceSource, byte experienceToAdd)
+    public void AddExperience(UnityEngine.Object experienceSource, float experienceToAdd)
     {
         float gainedExperience = experienceToAdd * GameData.InGameAttributes.ExperienceMultiplier;
